Resolves #9 Refactor to make unit tests possible
Previous versions of the code hardcoded dependencies towards types
Device, DeviceKind, InputCapability, Watcher, ... This made it
impossible to mock an implementation for testing purposes

With this refactoring, all the code is parameterized by a DeviceAccess
trait, which encapsulates all the lower-level APIs. To write unit
tests, we simply need to provide a default mockup implementation of
this trait.